ppc/AdeptusMechanicus/deploy/app/requirements.txt
regex-generator-lib
art
termcolor